Retire repositories after the thread that created them terminates
Fixes problems with leaking/missing arrangements with NUnit. NUnit reuses its test runner process and creates a new thread for each test run.
After the run completes, the thread terminates, but the test method repos remain usable and leak to the following test runs.
This happens because they are valid on all threads which is required for mocking async methods.